created universal markdown translatable editor
