+++ formwidgets/MLBlogMarkdown.php
@@ -1,8 +1,6 @@
 <?php namespace RainLab\Translate\FormWidgets;
 
-use RainLab\Blog\FormWidgets\BlogMarkdown;
 use RainLab\Blog\Models\Post;
-use RainLab\Translate\Models\Locale;
 
 /**
  * ML Blog Markdown
@@ -12,78 +10,12 @@ use RainLab\Translate\Models\Locale;
  * @package rainlab\translate
  * @author Rafał Soboń
  */
-class MLBlogMarkdown extends BlogMarkdown
+class MLBlogMarkdown extends MLMarkdownEditor
 {
-    use \RainLab\Translate\Traits\MLControl;
 
-    /**
-     * {@inheritDoc}
-     */
-    protected $defaultAlias = 'mlblogmarkdown';
-    public $originalViewPath;
-    public $originalAssetPath;
-
-    /**
-     * {@inheritDoc}
-     */
-    public function init()
-    {
+    public function init() {
         $this->actAsParent();
         parent::init();
-        $this->actAsParent(false);
-        $this->initLocale();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function render()
-    {
-        $this->actAsParent();
-        $parentContent = parent::render();
-        $this->actAsParent(false);
-
-        if (!$this->isAvailable) {
-            return $parentContent;
-        }
-
-        $this->vars['markdowneditor'] = $parentContent;
-        return $this->makePartial('mlblogmarkdown');
-    }
-
-    public function prepareVars()
-    {
-        parent::prepareVars();
-        $this->prepareLocaleVars();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    protected function loadAssets()
-    {
-        $this->actAsParent();
-        parent::loadAssets();
-        $this->actAsParent(false);
-
-        if (Locale::isAvailable()) {
-            $this->loadLocaleAssets();
-            $this->addJs('js/mlblogmarkdown.js');
-        }
-    }
-
-    private function actAsParent($switch = true)
-    {
-        if ($switch) {
-            $this->originalAssetPath = $this->assetPath;
-            $this->originalViewPath = $this->viewPath;
-            $this->assetPath = '/modules/backend/formwidgets/markdowneditor/assets';
-            $this->viewPath = base_path().'/modules/backend/formwidgets/markdowneditor/partials';
-        }
-        else {
-            $this->assetPath = $this->originalAssetPath;
-            $this->viewPath = $this->originalViewPath;
-        }
     }
 
     public function getSaveValue($value)
@@ -102,6 +34,12 @@ class MLBlogMarkdown extends BlogMarkdown
         }
 
         return array_get($localeData, $this->defaultLocale->code, $value);
+    }
+
+    private function actAsParent($switch = true)
+    {
+        $this->assetPath = "/plugins/rainlab/translate/formwidgets/mlmarkdowneditor/assets";
+        $this->viewPath = base_path()."/plugins/rainlab/translate/formwidgets/mlmarkdowneditor/partials";
     }
 
 }
